requirements.txt
# Requirements for DocuOps Agent
google-genai
python-dotenv
requests
chromadb
streamlit